mooved large function into own file
