+++ src/datasource.ts
@@ -11,10 +11,9 @@ import {
   TimeSeries,
   AppEvent,
   MutableDataFrame,
-  FieldType,
 } from '@grafana/data';
 import { BackendSrv, getBackendSrv, getTemplateSrv, SystemJS, TemplateSrv } from '@grafana/runtime';
-import { get, isEmpty, partition, reduce, uniq } from 'lodash';
+import { isEmpty, partition } from 'lodash';
 import {
   concurrent,
   datapointsPath,
@@ -31,6 +30,7 @@ import {
   convertItemsToTable,
   fetchRelationships,
 } from './cdf/client';
+import { createRelationshipsNode } from './cdf/relationships';
 import {
   AssetsFilterRequestParams,
   EventsFilterRequestParams,
@@ -49,8 +49,6 @@ import {
   EVENTS_PAGE_LIMIT,
   responseWarningEvent,
   EVENTS_LIMIT_WARNING,
-  nodeField,
-  edgeField,
 } from './constants';
 import { parse as parseQuery } from './parser/events-assets';
 import { formQueriesForExpression } from './parser/ts';
@@ -438,102 +436,12 @@ export default class CogniteDatasource extends DataSourceApi<
             this.connector
           );
           return !isEmpty(relationshipsList)
-            ? this.createRelationshipsNode(relationshipsList, refId)
+            ? createRelationshipsNode(relationshipsList, refId)
             : [];
         }
         return [];
       })
     ).then((res) => res[0]);
-  };
-
-  createRelationshipsNode = (relationshipsList, refId): Promise<MutableDataFrame[]> => {
-    const generateDetailKey = (key: string): string => ['detail__', key.trim().split(' ')].join('');
-
-    const allMetaKeysFromSourceAndTarget = reduce(
-      relationshipsList,
-      (previousValue, currentValue) => {
-        if (currentValue.source?.metadata) {
-          Object.keys(currentValue.source.metadata).map((key) => previousValue.push(key));
-        }
-        if (currentValue.target?.metadata) {
-          Object.keys(currentValue.target.metadata).map((key) => previousValue.push(key));
-        }
-        return uniq(previousValue);
-      },
-      []
-    );
-    const extendedFields = allMetaKeysFromSourceAndTarget.reduce((previousValue, currentValue) => {
-      return {
-        ...previousValue,
-        [generateDetailKey(currentValue)]: {
-          type: FieldType.string,
-          config: {
-            displayName: currentValue,
-          },
-        },
-      };
-    }, nodeField);
-    const edges = new MutableDataFrame({
-      name: 'edges',
-      fields: Object.keys(edgeField).map((key) => ({
-        ...edgeField[key],
-        name: key,
-      })),
-      meta: {
-        preferredVisualisationType: 'nodeGraph',
-      },
-      refId,
-    });
-    const nodes = new MutableDataFrame({
-      name: 'nodes',
-      fields: Object.keys(extendedFields).map((key) => ({
-        ...extendedFields[key],
-        name: key,
-      })),
-      meta: {
-        preferredVisualisationType: 'nodeGraph',
-      },
-      refId,
-    });
-    return relationshipsList.map(
-      ({ externalId, labels, sourceExternalId, targetExternalId, source, target }) => {
-        const { sourceMeta, targetMeta } = allMetaKeysFromSourceAndTarget.reduce((a, key) => {
-          const selector = generateDetailKey(key);
-          return {
-            sourceMeta: {
-              ...a.sourceMeta,
-              [selector]: get(source, `metadata.${key}`),
-            },
-            targetMeta: {
-              ...a.targetMeta,
-              [selector]: get(target, `metadata.${key}`),
-            },
-          };
-        }, {});
-        nodes.add({
-          id: sourceExternalId,
-          title: get(source, 'description'),
-          mainStat: get(source, 'name'),
-          ...sourceMeta,
-        });
-        nodes.add({
-          id: targetExternalId,
-          title: get(target, 'description'),
-          mainStat: get(target, 'name'),
-          ...targetMeta,
-        });
-        edges.add({
-          id: externalId,
-          source: sourceExternalId,
-          target: targetExternalId,
-          mainStat: labels
-            .map(({ externalId }) => externalId)
-            .join(', ')
-            .trim(),
-        });
-        return [nodes, edges];
-      }
-    )[relationshipsList.length - 1];
   };
 
   async checkLoginStatusApiKey() {
